Rename files and only run main.py
